+++ adapters/faqs/index.js
@@ -3,11 +3,12 @@ import { categoryAdapter } from '@/adapters/category';
 export const faqAdapter = ({ data }) => {
   if (data === null) return null;
   const { id, attributes } = data;
-  const { question, category, answer } = attributes;
+  const { question, category, answer, questionType } = attributes;
 
   return {
     id,
     question,
+    questionType: categoryAdapter(questionType),
     category: categoryAdapter(category),
     answer: answer === null ? '' : answer,
   };
@@ -15,7 +16,14 @@ export const faqAdapter = ({ data }) => {
 
 export const faqsAdapter = ({ data }) => {
   if (data === null) return [];
-  return data.map((faq) => {
-    return faqAdapter({ data: faq });
-  });
+  return data
+    .filter(({ attributes }) => attributes.category.data !== null)
+    .map((faq) => {
+      return faqAdapter({ data: faq });
+    });
+};
+
+export const faqsByTypeAdapter = ({ data }, categoryId) => {
+  if (data === null) return [];
+  return data.filter(({ category, questionType }) => category.id === categoryId && questionType !== null);
 };