+++ app/forgot-password/page.js
@@ -18,7 +18,7 @@ const ForgotPasswordPage = () => {
     <AuthWrapper
       title="Forgot your password?"
       subtitle="Enter your email address and you will receive an email with password reset link"
-      containerClass="w-3/4 -translate-y-14"
+      containerClass="md:px-48 -translate-y-14"
     >
       <ForgotPasswordForm />
       <NextLink href={ROUTES.LOGIN} className="inline-flex w-full justify-center text-blue text-xsm pt-2.5">