+++ app/reset-password/layout.js
@@ -4,10 +4,10 @@ import { ROUTES } from '@/lib';
 export default async function RootLayout(props) {
   const { children } = props;
   const navigation = {
-    placeholder: 'Already have an account?',
+    placeholder: 'Don’t have an account?',
     contrasted: true,
-    cta: 'Log in',
-    path: ROUTES.LOGIN,
+    cta: 'Registration',
+    path: ROUTES.SIGNUP,
   };
   return <AuthLayout navigation={navigation}>{children}</AuthLayout>;
 }