+++ app/success/page.js
@@ -1,6 +1,5 @@
 import { metaData } from '@/adapters/metaData';
 import { LinkAsButton, NextLink } from '@/elements';
-import { ROUTES } from '@/lib';
 import { AuthWrapper } from '@/modules';
 
 export function generateMetadata() {
@@ -15,7 +14,7 @@ export function generateMetadata() {
 
 const SuccessfulVerifPage = () => {
   return (
-    <AuthWrapper title="Identity Verification Successful!" containerClass="w-3/4 -translate-y-14 min-w-[450px] pt-20">
+    <AuthWrapper title="Identity Verification Successful!" containerClass="md:px-28 min-w-[450px]">
       <div className="grid gap-2 my-5 text-xsm">
         <p>
           Congratulations! You have successfully passed our identity verification procedure. We take the security of our
@@ -36,8 +35,8 @@ const SuccessfulVerifPage = () => {
         </p>
       </div>
       <p className="my-5 text-xsm">Thank you for choosing our platform!</p>
-      <LinkAsButton href={ROUTES.LOGIN} buttonProps={{ size: 'large', variant: 'primary' }}>
-        Log in
+      <LinkAsButton href="/" buttonProps={{ size: 'large', variant: 'primary' }}>
+        Return to Homepage
       </LinkAsButton>
     </AuthWrapper>
   );