+++ components/elements/Button/index.js
@@ -8,6 +8,7 @@ import { getButtonClassNames } from '@/utils/helpers';
 const Button = ({
   buttonProps: { icon = {}, iconContainerStyles, helperText, text, variant, size },
   customStyles,
+  customStylesFromWrap,
   disabled = false,
   type = 'button',
   onClick,
@@ -16,10 +17,10 @@ const Button = ({
   const { before, after } = icon;
   const buttonClassNames = getButtonClassNames(variant, size);
   return (
-    <div className="flex flex-col justify-center items-center">
+    <div className={`flex flex-col justify-center items-center ${customStylesFromWrap}`}>
       <button
         className={classnames(
-          'text-xsm px-3.5 py-2.5 rounded-md flex items-center justify-center',
+          'text-xsm px-3.5 py-2.5 rounded-md flex items-center justify-center group',
           buttonClassNames,
           disabled && 'opacity-50 pointer-events-none',
           customStyles