+++ components/modules/DeactivateAccountForm/index.js
@@ -6,7 +6,7 @@ import * as yup from 'yup';
 
 import { DeactivateAccountFormPropTypes } from '@/lib/types';
 
-import { FormManager } from '@/common';
+import { ModalFormManager } from '@/common';
 import { Input, Title } from '@/elements';
 import { useHookFormParams } from '@/utils/hooks';
 
@@ -16,7 +16,7 @@ const state = {
 
 const schema = yup.object({ password: yup.string().required('Required field') });
 
-const DeactivateAccountForm = ({ title }) => {
+const DeactivateAccountForm = ({ title, closeModal }) => {
   const methods = useHookFormParams({ state, schema });
 
   const {
@@ -29,29 +29,33 @@ const DeactivateAccountForm = ({ title }) => {
   };
 
   return (
-    <FormProvider {...methods}>
-      <FormManager
-        className="max-w-sm"
-        submitAction={onSubmit}
-        submitButton={{ text: 'Deactivate Account', variant: 'delete', size: 'large' }}
-      >
-        <div className="text-black flex flex-col gap-2.5">
-          <Title level="3" className="text-lg font-bold">
-            {title}
-          </Title>
-          <p className="text-xsm">
-            <span className="font-bold">
-              We will deactivate all your tankers and your account will become inactive.
-            </span>
-            But if you want to reactivate your account, then after logging in, we will automatically reactivate your
-            account.
-          </p>
-        </div>
+    <div className="w-[356px]">
+      <FormProvider {...methods}>
+        <ModalFormManager
+          onClose={closeModal}
+          className="max-w-sm "
+          submitAction={onSubmit}
+          specialStyle
+          submitButton={{ text: 'Deactivate Account', variant: 'delete', size: 'large' }}
+        >
+          <div className="text-black flex flex-col gap-2.5">
+            <Title level="3" className="text-lg font-bold">
+              {title}
+            </Title>
+            <p className="text-xsm">
+              <span className="font-bold">
+                We will deactivate all your tankers and your account will become inactive.
+              </span>
+              But if you want to reactivate your account, then after logging in, we will automatically reactivate your
+              account.
+            </p>
+          </div>
 
-        <p className="font-bold text-black">Please enter your password to deactivate account</p>
-        <Input {...register('password')} type="password" label="password" error={errors.password?.message} />
-      </FormManager>
-    </FormProvider>
+          <p className="font-bold text-black">Please enter your password to deactivate account</p>
+          <Input {...register('password')} type="password" label="password" error={errors.password?.message} />
+        </ModalFormManager>
+      </FormProvider>
+    </div>
   );
 };