+++ components/modules/Notification/index.js
@@ -7,7 +7,7 @@ const Notification = ({ numberOfNotifications }) => {
   return (
     <div className="relative">
       <HoverableIcon icon={<NotificationSVG />} />
-      <div className="absolute -top-1 -right-2 w-5 h-5 rounded-[50%] bg-blue text-[10px] font-bold text-white flex items-center justify-center">
+      <div className="absolute border border-white border-solid -top-1 -right-2 px-0.5 h-5 min-w-[20px] rounded-base bg-blue text-xxs font-bold text-white flex items-center justify-center">
         {numberOfNotifications}
       </div>
     </div>