+++ components/modules/OwnerRegistrationForm/index.js
@@ -26,7 +26,7 @@ import {
   TermsAndConditions,
 } from '@/units';
 import { resetForm } from '@/utils/helpers';
-import { errorToast, successToast, useHookFormParams } from '@/utils/hooks';
+import { errorToast, redirectAfterToast, useHookFormParams } from '@/utils/hooks';
 
 const OwnerRegistrationForm = () => {
   const [sameAddress, setSameAddress] = useState(false);
@@ -51,8 +51,8 @@ const OwnerRegistrationForm = () => {
   const onSubmit = async (formData) => {
     const { data, error } = await ownerSignUp({ data: formData });
     if (data) {
-      successToast(data.message);
-      resetForm(methods);
+      resetForm(methods, '');
+      redirectAfterToast(data.message, '/');
     } else if (error) {
       errorToast(error.message, error.errors.join('\n'));
     }