+++ components/units/index.js
@@ -4,6 +4,7 @@ import AccordionHeader from '@/units/AccordionHeader';
 import AddressDetails from '@/units/AddressDetailsForm';
 import CargoesSlotsDetails from '@/units/CargoesSlotsDetailsForm';
 import CargoIdTooltip from '@/units/CargoIdTooltip';
+import ChartererInformationContent from '@/units/ChartererInformationContent';
 import Comment from '@/units/Comment';
 import CommercialOfferTerms from '@/units/CommercialOfferTerms';
 import CompanyAddresses from '@/units/CompanyAddressesForm';
@@ -11,6 +12,7 @@ import CompanyDetails from '@/units/CompanyDetailsForm';
 import ComplexPagination from '@/units/ComplexPagination';
 import COTTabContent from '@/units/COTTabContent';
 import Countdown from '@/units/Countdown';
+import CounterofferForm from '@/units/CounterofferForm';
 import DateDetailsForm from '@/units/DateDetailsForm';
 import DeactivateTankerForm from '@/units/DeactivateTankerForm';
 import EditDateForm from '@/units/EditDateForm';
@@ -18,6 +20,7 @@ import EditPortForm from '@/units/EditPortForm';
 import ExpandableCard from '@/units/ExpandableCard';
 import ExpandableRowFooter from '@/units/ExpandableRowFooter';
 import ExpandableRowHeader from '@/units/ExpandableRowHeader';
+import UploadForm from '@/units/FileUpload';
 import FooterNavBlock from '@/units/FooterNavBlock';
 import IconWrapper from '@/units/IconWrapper';
 import LegalNavigation from '@/units/LegalNavigation';
@@ -28,6 +31,7 @@ import OfferAcceptModalContent from '@/units/OfferAcceptModalContent';
 import OfferDeclineForm from '@/units/OfferDeclineForm';
 import OfferForm from '@/units/OfferForm';
 import PaginationComponent from '@/units/PaginationComponent';
+import PartyItem from '@/units/PartyItem';
 import PasswordValidation from '@/units/PasswordValidationForm';
 import PersonalDetails from '@/units/PersonalDetailsForm';
 import PortDetailsForm from '@/units/PortDetailsForm';
@@ -38,6 +42,8 @@ import SocialNetworks from '@/units/SocialNetworks';
 import Step from '@/units/Step';
 import Tabs from '@/units/Tabs';
 import TabsAsLinks from '@/units/TabsAsLinks';
+import TabsVertical from '@/units/TabsVertical';
+import TankerInformationContent from '@/units/TankerInformationContent';
 import TankerSlotsDetails from '@/units/TankerSlotsDetailsForm';
 import TermsAndConditions from '@/units/TermsAndConditionsForm';
 import ToggleRows from '@/units/ToggleRows';
@@ -91,5 +97,11 @@ export {
   SocialNetworks,
   ReactivateTankerForm,
   LegalNavigation,
+  PartyItem,
+  UploadForm,
+  TabsVertical,
   CargoIdTooltip,
+  CounterofferForm,
+  ChartererInformationContent,
+  TankerInformationContent,
 };