+++ pages/api/auth/[...nextauth].js
@@ -1,7 +1,6 @@
 /* eslint-disable import/no-extraneous-dependencies */
 import NextAuth from 'next-auth';
 
-export const authOptions = {
-  providers: [],
-};
+import { authOptions } from '@/lib/auth';
+
 export default NextAuth(authOptions);