+++ services/user/index.js
@@ -29,8 +29,13 @@ export async function resetPassword({ data }) {
 export async function ownerSignUp({ data }) {
   const body = ownerSignUpAdapter({ data });
   const response = await postData(`auth/signup?type=owner`, body);
+
   return {
     ...response,
+    data: {
+      message:
+        !response.error && 'To confirm registration, follow the link that was sent to the email address, you provided',
+    },
   };
 }
 
@@ -39,6 +44,10 @@ export async function chartererSignUp({ data }) {
   const response = await postData(`auth/signup?type=charterer`, body);
   return {
     ...response,
+    data: {
+      message:
+        !response.error && 'To confirm registration, follow the link that was sent to the email address, you provided',
+    },
   };
 }
 
@@ -111,6 +120,13 @@ export async function getUserNegotiating() {
 
 export async function getUserDetails() {
   const response = await getData(`account/user-info`);
+  return {
+    ...response,
+  };
+}
+
+export async function getUserOnSubs() {
+  const response = await getData(`account/on-subs`);
   return {
     ...response,
   };